YQL-19747 Extract syntax module
