+++ yql/essentials/sql/v1/complete/syntax/sql_syntax.cpp
