+++ yql/essentials/sql/v1/complete/syntax/ya.make
@@ -1,0 +1,26 @@
+LIBRARY()
+
+SRCS(
+    sql_antlr4.cpp
+    sql_context.cpp
+    sql_syntax.cpp
+)
+
+ADDINCL(
+    yql/essentials/sql/v1/complete
+)
+
+PEERDIR(
+    yql/essentials/core/issue
+
+    yql/essentials/parser/antlr_ast/gen/v1_ansi_antlr4
+    yql/essentials/parser/antlr_ast/gen/v1_antlr4
+
+    yql/essentials/sql/settings
+    yql/essentials/sql/v1/lexer
+
+    # TODO(YQL-19747): Replace with the sql/v1/reflect to get keywords
+    yql/essentials/sql/v1/format
+)
+
+END()
